# docs/stream-tide.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "stream-tide",
    "description": "",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/users/register": {
      "post": {
        "summary": "Register new user",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "User"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "userName": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "userName",
                  "password"
                ],
                "x-apidog-orders": [
                  "firstName",
                  "lastName",
                  "email",
                  "userName",
                  "password"
                ],
                "x-apidog-ignore-properties": []
              },
              "example": {
                "firstName": "Jane",
                "lastName": "Doe",
                "email": "shohanur1233@gmail.com",
                "userName": "shohanurriidofdfdesss4",
                "password": "SecurePass1!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "userName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "firstName",
                    "lastName",
                    "userName",
                    "email",
                    "accessToken",
                    "refreshToken"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "id",
                    "firstName",
                    "lastName",
                    "userName",
                    "email",
                    "accessToken",
                    "refreshToken"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9541952-run",
        "security": []
      }
    },
    "/api/v1/users/login": {
      "post": {
        "summary": "User login",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "User"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "x-apidog-orders": [
                  "email",
                  "password"
                ],
                "x-apidog-ignore-properties": []
              },
              "example": {
                "email": "jane.doe@example434.com",
                "password": "SecurePass1!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "accessToken",
                        "refreshToken"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "id",
                        "accessToken",
                        "refreshToken"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9541974-run",
        "security": []
      }
    },
    "/api/v1/users/logout-all": {
      "delete": {
        "summary": "Logout all ",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "User"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9541976-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "summary": "Get user  list",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": false,
            "example": "2",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "firstName": {
                            "type": "string"
                          },
                          "lastName": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "firstName",
                          "lastName"
                        ],
                        "x-apidog-ignore-properties": [],
                        "x-apidog-orders": [
                          "id",
                          "firstName",
                          "lastName"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "totalItems": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "currentPage",
                        "pageSize",
                        "totalItems",
                        "totalPages"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "currentPage",
                        "pageSize",
                        "totalItems",
                        "totalPages"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data",
                    "pagination"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9541979-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "summary": "Get single details",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "userName": {
                          "type": "string"
                        },
                        "image": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "firstName",
                        "lastName",
                        "email",
                        "userName",
                        "image",
                        "createdAt",
                        "role"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "id",
                        "firstName",
                        "lastName",
                        "email",
                        "userName",
                        "image",
                        "createdAt",
                        "role"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9541980-run",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "summary": "Update user details",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "Jhon"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe "
                  },
                  "file": {
                    "type": "string",
                    "example": "file:///home/shohanur/Temp/test-images/Wiki_Test_Image.jpg",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "userName": {
                          "type": "string"
                        },
                        "imageUrl": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "firstName",
                        "lastName",
                        "email",
                        "userName",
                        "imageUrl"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "id",
                        "firstName",
                        "lastName",
                        "email",
                        "userName",
                        "imageUrl"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9541981-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/refresh": {
      "post": {
        "summary": "Refresh access token",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "User"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": [
                  "refreshToken"
                ],
                "x-apidog-orders": [
                  "refreshToken"
                ],
                "x-apidog-ignore-properties": []
              },
              "example": {
                "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NiwiZW1haWwiOiJqYW5lLmRvZUBleGFtcGxlNDM0LmNvbSIsInJvbGUiOiJhZG1pbiIsInJlZnJlc2hUb2tlbklkIjoiNjpiOTgwZTdlYy1lM2VmLTQ1NzctYmE3MC0xNzU0OWRjMDAzNTMiLCJpYXQiOjE3MjQ1MjUzOTIsImV4cCI6MTcyNDUyNjI5Mn0.dMYYosrYK7on3DbBsLVSztfcoE5HSXR-735dA9VgVhg"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "accessToken",
                        "refreshToken"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "accessToken",
                        "refreshToken"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9541989-run",
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Develop Env"
    }
  ]
}